Added task images and solution screenshot
Problem Statement:

Basically, you have data of two different events, Lola and Hula of the years 2014-2017.
First, I want to have two boxes to choose which event and which year to be displayed.

Next, the scores of behavioural intentions, satisfaction and line-up should be displayed in 3 different boxes. All can have a score between 1-100. In the screenshot, you see a blue line in each box and it represents the score / 100. Next, I would like to see the score in blue. I also would like to see the absolute and relative change compared to the previous period indicated with a green arrow for a positive change and red for negative. If there is no change, everything should be empty and there should a grey circle.